big update - Partener Group - levels - vote add shop group

+++ src/main/java/ru/skidoz/aop/repo/CashbackCacheRepository.java
@@ -1,6 +1,6 @@
 package ru.skidoz.aop.repo;
 
-import java.math.BigDecimal;
+
 import java.util.List;
 import ru.skidoz.model.entity.ActionTypeEnum;
 import ru.skidoz.model.pojo.side.Cashback;
@@ -25,7 +25,7 @@ public interface CashbackCacheRepository extends JpaRepositoryTest<Cashback, Int
 
     List<Cashback> findAllByUser_IdAndShop_IdAndAction_Type(Integer  buyerId, Integer  shopId, ActionTypeEnum actionTypeEnum);
 
-    BigDecimal purchaseSumByUserAndShopAndAction_Type(Integer  buyerId, Integer  shopId, ActionTypeEnum actionTypeEnum);
+    Integer purchaseSumByUserAndShopAndAction_Type(Integer  buyerId, Integer  shopId, ActionTypeEnum actionTypeEnum);
 
     void deleteById(Integer id, Integer purchaseId, Integer actionId, ActionTypeEnum actionTypeEnum);
